[RT#89770] Patch to fix error reporting if command killed by signal
Currently if the command is killed by a signal, IPC::Cmd gives an error message containing '%d' and '%s' instead of the information.

% perl -MIPC::Cmd -E '($ok, $err) = IPC::Cmd::run(command => [ qw(perl -E dump) ]); say $err'
'perl -E dump' died with signal %d, %s coredump

This is because the error string contains printf() style placeholders, which are not understood by loc(), the function used to generate localized error messages.  Fix it like this:

With this fix, the correct error is given:

% perl -MIPC::Cmd -E '($ok, $err) = IPC::Cmd::run(command => [ qw(perl -E dump) ]); say $err'
'perl -E dump' died with signal 6, with coredump

Signed-off-by: Chris 'BinGOs' Williams <chris@bingosnet.co.uk>